core/app.py
import sys
import os
from pathlib import Path
from loguru import logger
from dotenv import load_dotenv

# Add project root to Python path
project_root = Path(__file__).parent.parent
sys.path.append(str(project_root))

# Load environment variables
load_dotenv()

# Configure logging
logger.add(
    project_root / "logs" / "app.log",
    rotation="500 MB",
    retention="10 days",
    level="INFO"
)

class HiccupApp:
    def __init__(self):
        self.config = {}
        self.screenpipe = None
        self.terminator = None
        
    def initialize(self):
        """Initialize the application components"""
        try:
            # Initialize components here
            logger.info("Initializing Hiccup application...")
            # TODO: Initialize ScreenPipe and Terminator components
            return True
        except Exception as e:
            logger.error(f"Failed to initialize application: {str(e)}")
            return False
            
    def run(self):
        """Main application loop"""
        if not self.initialize():
            logger.error("Failed to start application")
            return
            
        try:
            logger.info("Starting Hiccup application...")
            # TODO: Implement main application loop
            pass
        except KeyboardInterrupt:
            logger.info("Application stopped by user")
        except Exception as e:
            logger.error(f"Application error: {str(e)}")
        finally:
            self.cleanup()
            
    def cleanup(self):
        """Cleanup resources"""
        logger.info("Cleaning up resources...")
        # TODO: Implement cleanup logic

if __name__ == "__main__":
    app = HiccupApp()
    app.run() 