requirements.txt
pyautogui==0.9.54
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
python-dotenv==1.0.0
loguru==0.7.2
pynput==1.7.6
keyboard==0.13.5
mouse==0.7.1 